cloud_functions/google_trends_scraper/requirements.txt
functions-framework>=3.0.0
google-cloud-bigquery>=3.0.0
google-cloud-storage>=2.0.0
pandas>=1.3.5
pytrends>=4.9.0
